+++ src/desktop/Studio/types/studioType.ts
@@ -1,0 +1,11 @@
+export interface StudioInfoProps {
+  info: {
+    studio: string;
+    advisor: string;
+    description: string;
+  };
+}
+
+export interface TotalWorksProps {
+  id: number;
+}
